+++ hoomd/hpmc/IntegratorHPMCMono.h
@@ -543,7 +543,7 @@ template<class Shape> void IntegratorHPMCMono<Shape>::update(uint64_t timestep)
                     continue;
                     }
 
-                move_translate(pos_i, rng_i, h_d.data[typ_i], ndim);
+                move_translate(pos_i, rng_i, h_d.data[typ_i], this->getTranslateMoveDimensions());
 
 #ifdef ENABLE_MPI
                 if (m_sysdef->isDomainDecomposed())